Move option parsion code into separate module
